refactor(lostpassword): change method name to recovery and reset
